feat: Add Family Book feature
Introduces the ability to view the family lineage and past ancestors through a new "Family Book" feature. This includes adding the `Ancestor` type and a modal component for displaying ancestor information and history. The `CharacterSheet` and `App.tsx` are updated to support opening this new modal. A `LockClosedIcon` has also been added.